add unit test for frost/helper.go
